+++ base/src/main/java/proguard/classfile/io/kotlin/KotlinMetadataWriter.java
@@ -1808,8 +1808,6 @@ implements ClassVisitor,
     {
         Set<Flag> flagSet = new HashSet<>();
 
-        flagSet.addAll(convertCommonFlags(flags.common));
-
         if (flags.isNullable)          flagSet.add(Flag.Type.IS_NULLABLE);
         if (flags.isSuspend)           flagSet.add(Flag.Type.IS_SUSPEND);
         if (flags.isDefinitelyNonNull) flagSet.add(Flag.Type.IS_DEFINITELY_NON_NULL);