+++ base/src/test/kotlin/proguard/classfile/kotlin/flags/KotlinTypeFlagsTest.kt
@@ -191,6 +191,112 @@ class KotlinTypeFlagsTest : FreeSpec({
             }
         }
     }
+
+    "Given a type argument with name annotation" - {
+        val clazz = ClassPoolBuilder.fromSource(
+            KotlinSource(
+                "Test.kt",
+                """
+                val namedParam: (test: String) -> Unit = {}
+                """
+            )
+        ).programClassPool.getClass("TestKt")
+
+        "Then the flags should be initialized correctly" {
+            clazz.accept(
+                ReferencedKotlinMetadataVisitor(
+                    AllTypeVisitor(
+                        KotlinTypeFilter(
+                            {
+                                it.className == "kotlin/Function1" && it.typeArguments.size == 2
+                            },
+                            { _, kotlinTypeMetadata ->
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isNullable shouldBe false
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isSuspend shouldBe false
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isDefinitelyNonNull shouldBe false
+
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.common?.hasAnnotations shouldBe true
+                            }
+                        )
+                    )
+                )
+            )
+        }
+
+        "Then the flags should be written and re-initialized correctly" {
+            clazz.accept(
+                ReWritingMetadataVisitor(
+                    AllTypeVisitor(
+                        KotlinTypeFilter(
+                            {
+                                it.className == "kotlin/Function1" && it.typeArguments.size == 2
+                            },
+                            { _, kotlinTypeMetadata ->
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isNullable shouldBe false
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isSuspend shouldBe false
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isDefinitelyNonNull shouldBe false
+
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.common?.hasAnnotations shouldBe true
+                            }
+                        )
+                    )
+                )
+            )
+        }
+    }
+
+    "Given a nullable type argument" - {
+        val clazz = ClassPoolBuilder.fromSource(
+            KotlinSource(
+                "Test.kt",
+                """
+                val namedParam: (String?) -> Unit = {}
+                """
+            )
+        ).programClassPool.getClass("TestKt")
+
+        "Then the flags should be initialized correctly" {
+            clazz.accept(
+                ReferencedKotlinMetadataVisitor(
+                    AllTypeVisitor(
+                        KotlinTypeFilter(
+                            {
+                                it.className == "kotlin/Function1" && it.typeArguments.size == 2
+                            },
+                            { _, kotlinTypeMetadata ->
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isNullable shouldBe true
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isSuspend shouldBe false
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isDefinitelyNonNull shouldBe false
+
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.common?.hasAnnotations shouldBe false
+                            }
+                        )
+                    )
+                )
+            )
+        }
+
+        "Then the flags should be written and re-initialized correctly" {
+            clazz.accept(
+                ReWritingMetadataVisitor(
+                    AllTypeVisitor(
+                        KotlinTypeFilter(
+                            {
+                                it.className == "kotlin/Function1" && it.typeArguments.size == 2
+                            },
+                            { _, kotlinTypeMetadata ->
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isNullable shouldBe true
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isSuspend shouldBe false
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.isDefinitelyNonNull shouldBe false
+
+                                kotlinTypeMetadata.typeArguments?.get(0)?.flags?.common?.hasAnnotations shouldBe false
+                            }
+                        )
+                    )
+                )
+            )
+        }
+    }
 })
 
 private fun createVisitor(className: String, typeVisitor: KotlinTypeVisitor): KotlinMetadataVisitor =