+++ docs/md/releasenotes.md
@@ -4,6 +4,7 @@
 
 - Enable fix previously behind system property: fix `TypedReferenceValue.generalize()` not setting `mayBeExtension` to true when generalizing to common parent type.
 - Avoid printing `PartialEvaluator` messages when an `ExcessiveComplexityException` occurs.
+- Fix incorrect writing of flags for type parameters with name annotations.
 
 ### Improved